UserManagerMicroservice/requirements.txt
flask
grpcio
grpcio-tools
mysql-connector-python
protobuf
redis